fix: move to utility
