+++ src/v0/destinations/campaign_manager/transform.js
@@ -18,36 +18,13 @@ const {
   EncryptionSource,
 } = require('./config');
 
+const { convertToMicroseconds } = require('./util');
+
 const { InstrumentationError } = require('../../util/errorTypes');
 const { JSON_MIME_TYPE } = require('../../util/constant');
 
 function isEmptyObject(obj) {
   return Object.keys(obj).length === 0 && obj.constructor === Object;
-}
-
-function convertToMicroseconds(input) {
-  const timestamp = Date.parse(input);
-
-  if (!Number.isNaN(timestamp)) {
-    // If the input is a valid date string, timestamp will be a number
-    if (input.includes('Z')) {
-      // ISO 8601 date string with milliseconds
-      return timestamp * 1000;
-    }
-    // to handle case of "2022-11-17T00:22:02.903+05:30" strings
-    return timestamp.toString().length <= 13 ? timestamp * 1000 : timestamp * 1000000;
-  }
-
-  if (/^\d+$/.test(input)) {
-    // If the input is a numeric string (assume microseconds or milliseconds)
-    if (input.length <= 13) {
-      // Length less than or equal to 13 indicates milliseconds
-      return parseInt(input, 10) * 1000;
-    }
-    // Otherwise, assume microseconds
-    return parseInt(input, 10);
-  }
-  return timestamp;
 }
 
 // build final response
